move code to applications.go service
